add: factories for entities; in progress: CustomerRepositoryImpl;

+++ sql/create_tables_script.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS scooter_rents(
-scooter_rent_id BIGINT PRIMARY KEY NOT NULL,
+scooter_rent_id SERIAL PRIMARY KEY NOT NULL,
 is_ongoing BOOLEAN NOT NULL,
 period TIME NOT NULL,
 customer_id BIGINT NOT NULL,
@@ -9,7 +9,7 @@ scooter_id BIGINT NOT NULL
 CREATE TYPE scooter_type AS ENUM ('long_range', 'urban', 'folding');
 
 CREATE TABLE IF NOT EXISTS customers(
-customer_id BIGINT PRIMARY KEY NOT NULL,
+customer_id SERIAL PRIMARY KEY NOT NULL,
 name VARCHAR(40),
 surname VARCHAR(40),
 email VARCHAR(40) UNIQUE,
@@ -18,7 +18,7 @@ scooter_rent_id BIGINT
 );
 
 CREATE TABLE IF NOT EXISTS Scooters(
-scooter_id BIGINT PRIMARY KEY NOT NULL,
+scooter_id SERIAL PRIMARY KEY NOT NULL,
 available BOOLEAN NOT NULL,
 type scooter_type NOT NULL,
 battery_level INT CHECK (battery_level>=0 AND battery_level<=100),
